修改 使用uint作为Id类型 修改 Response的StatusCode类型改为int

+++ src/controller/user_controller.go
@@ -37,7 +37,7 @@ func UserLogin(c *gin.Context) {
 
 func UserInfo(c *gin.Context) {
 	userIdStr, _ := c.Get("user_id")
-	userId := userIdStr.(int64)
+	userId := userIdStr.(uint)
 
 	userInfoResponse, err := service.UserInfo(userId)
